+++ CHANGELOG.md
@@ -2,4 +2,4 @@
 
 ## src/popup
 
-- added more to the add button component regarding the form
+- added basic form validation